Add backend dependencies and environment config
Created new `package.json`, `http-client.env.json` and `pnpm-lock.yaml` to add backend dependencies and establish a development environment that points to `localhost:7070`. This is essential for setting up the backend application. The `http-client.env.json` file defines the development URL for server side. The `package.json` and `pnpm-lock-yaml` file lists out all the dependencies and their specific versions which the application requires to work. This ensures consistency when others work on the project or when deploying for production so that everyone uses the exact same versions of each package or library.